+++ src/Core/VpnHood.Core.Tunneling/Proxies/ProxyManagerOptions.cs
@@ -10,7 +10,6 @@ public class ProxyManagerOptions
     public required int MaxUdpClientCount { get; init; }
     public required int MaxPingClientCount { get; init; }
     public required int PacketQueueCapacity { get; init; }
-    public required bool UseUdpProxy2 { get; init; }
     public required TransferBufferSize? UdpBufferSize { get; init; }
     public required bool IsPingSupported { get; init; }
     public required IPacketProxyCallbacks? PacketProxyCallbacks { get; init; }